rearrange polars compile module
